+++ reactivesocket-aeron/build.gradle
@@ -1,3 +1,4 @@
 dependencies {
     compile 'io.aeron:aeron-all:0.9.5'
+    testCompile project(':reactivesocket-test')
 }
