+++ reactivesocket-jsr-356/src/test/java/io/reactivesocket/javax/websocket/PongEndpoint.java
@@ -18,6 +18,7 @@ package io.reactivesocket.javax.websocket;
 import io.reactivesocket.Payload;
 import io.reactivesocket.RequestHandler;
 import io.reactivesocket.javax.websocket.server.ReactiveSocketWebSocketServer;
+import io.reactivesocket.test.TestUtil;
 import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
 import rx.Observable;
