+++ reactivesocket-local/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+  testCompile project(':reactivesocket-test')
+}
