+++ reactivesocket-local/src/test/java/io/reactivesocket/local/ClientServerTest.java
@@ -21,6 +21,7 @@ import io.reactivesocket.Payload;
 import io.reactivesocket.ReactiveSocket;
 import io.reactivesocket.RequestHandler;
 import io.reactivesocket.exceptions.SetupException;
+import io.reactivesocket.test.TestUtil;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.reactivestreams.Publisher;
