+++ reactivesocket-netty/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
     compile 'io.netty:netty-handler:4.1.0.CR7'
     compile 'io.netty:netty-codec-http:4.1.0.CR7'
+    testCompile project(':reactivesocket-test')
 }
 
 task echoServer(type: JavaExec) {
