+++ reactivesocket-netty/src/test/java/io/reactivesocket/netty/websocket/Pong.java
@@ -29,8 +29,8 @@ import io.netty.handler.logging.LogLevel;
 import io.netty.handler.logging.LoggingHandler;
 import io.reactivesocket.Payload;
 import io.reactivesocket.RequestHandler;
-import io.reactivesocket.netty.TestUtil;
 import io.reactivesocket.netty.websocket.server.ReactiveSocketServerHandler;
+import io.reactivesocket.test.TestUtil;
 import org.reactivestreams.Publisher;
 import org.reactivestreams.Subscriber;
 import rx.Observable;
