+++ reactivesocket-test/src/main/java/io/reactivesocket/test/TestUtil.java
@@ -13,9 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.reactivesocket.netty;
+package io.reactivesocket.test;
 
-import io.netty.buffer.ByteBuf;
 import io.reactivesocket.Frame;
 import io.reactivesocket.FrameType;
 import io.reactivesocket.Payload;
@@ -57,8 +56,10 @@ public class TestUtil
         return new PayloadImpl(data, metadata);
     }
 
-    public static String byteToString(final ByteBuffer byteBuffer)
+    public static String byteToString(ByteBuffer byteBuffer)
     {
+        byteBuffer = byteBuffer.duplicate();
+
         final byte[] bytes = new byte[byteBuffer.remaining()];
         byteBuffer.get(bytes);
         return new String(bytes, Charset.forName("UTF-8"));
@@ -119,42 +120,5 @@ public class TestUtil
         {
             return metadata;
         }
-    }
-
-    public static String byteBufToString(ByteBuf buf) {
-        byte[] bytes = new byte[buf.readableBytes()];
-        int readerIndex = buf.readerIndex();
-        buf.getBytes(readerIndex, bytes);
-        buf.readerIndex(readerIndex);
-
-        StringBuilder result = new StringBuilder();
-        StringBuilder ascii = new StringBuilder();
-        int i = 0;
-        for (i=0; i<bytes.length; i++) {
-            byte b = bytes[i];
-            result.append(String.format("%02X ", b));
-
-            if (32 <= b && b < 127) {
-                ascii.append((char)b);
-            } else {
-                ascii.append('.');
-            }
-
-            if ((i+1) % 16 == 0) {
-                result.append("   ");
-                result.append(ascii);
-                result.append('\n');
-                ascii = new StringBuilder();
-            }
-        }
-        if ((bytes.length - 1) % 16 != 0) {
-            int x = 16 - ((bytes.length - 1) % 16);
-            StringBuilder padding = new StringBuilder();
-            for (int j=0 ; j<x; j++) {
-                result.append("   ");
-            }
-            result.append(ascii);
-        }
-        return result.toString();
     }
 }
